+++ mutable_tree_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
 	db "github.com/tendermint/tm-db"
 )
 
@@ -189,8 +190,8 @@ func TestFlushVersion_Empty(t *testing.T) {
 }
 
 func TestDelete(t *testing.T) {
-	memDb := db.NewMemDB()
-	tree, err := NewMutableTree(memDb, 0)
+	memDB := db.NewMemDB()
+	tree, err := NewMutableTree(memDB, 0)
 	require.NoError(t, err)
 
 	tree.set([]byte("k1"), []byte("Fred"))
@@ -205,7 +206,7 @@ func TestDelete(t *testing.T) {
 	require.Nil(t, k1Value)
 
 	key := tree.ndb.rootKey(version)
-	err = memDb.Set(key, hash)
+	err = memDB.Set(key, hash)
 	require.NoError(t, err)
 	tree.versions[version] = true
 
@@ -215,8 +216,8 @@ func TestDelete(t *testing.T) {
 }
 
 func TestTraverse(t *testing.T) {
-	memDb := db.NewMemDB()
-	tree, err := NewMutableTree(memDb, 0)
+	memDB := db.NewMemDB()
+	tree, err := NewMutableTree(memDB, 0)
 	require.NoError(t, err)
 
 	for i := 0; i < 6; i++ {