+++ nodedb.go
@@ -7,8 +7,8 @@ import (
 	"sort"
 	"sync"
 
+	lru "github.com/hashicorp/golang-lru"
 	"github.com/pkg/errors"
-
 	"github.com/tendermint/tendermint/crypto/tmhash"
 	dbm "github.com/tendermint/tm-db"
 )
@@ -32,6 +32,8 @@ var (
 
 	// Root nodes are indexed separately by their version
 	rootKeyFormat = NewKeyFormat('r', int64Size) // r<version>
+
+	metadataKeyFormat = NewKeyFormat('m', int64Size) // m<version>
 )
 
 type nodeDB struct {
@@ -47,13 +49,21 @@ type nodeDB struct {
 	nodeCache      map[string]*list.Element // Node cache.
 	nodeCacheSize  int                      // Node cache size limit in elements.
 	nodeCacheQueue *list.List               // LRU queue of cache elements. Used for deletion.
+
+	vmCache *lru.Cache // LRU cache of version metadata
 }
 
 func newNodeDB(snapshotDB dbm.DB, recentDB dbm.DB, cacheSize int, opts *Options) *nodeDB {
 	if opts == nil {
 		opts = DefaultOptions()
 	}
-	ndb := &nodeDB{
+
+	vmCache, err := lru.New(20000)
+	if err != nil {
+		panic(fmt.Errorf("failed to create metadata cache: %w", err))
+	}
+
+	return &nodeDB{
 		snapshotDB:     snapshotDB,
 		recentDB:       recentDB,
 		snapshotBatch:  snapshotDB.NewBatch(),
@@ -64,12 +74,82 @@ func newNodeDB(snapshotDB dbm.DB, recentDB dbm.DB, cacheSize int, opts *Options)
 		nodeCacheSize:  cacheSize,
 		nodeCacheQueue: list.New(),
 		versionReaders: make(map[int64]uint32, 8),
+		vmCache:        vmCache,
 	}
-	return ndb
 }
 
-func (ndb *nodeDB) isSnapshotVersion(version int64) bool {
-	return ndb.opts.KeepEvery != 0 && version%ndb.opts.KeepEvery == 0
+// GetVersionMetadata returns a reference to VersionMetadata for a given version.
+// The object is first checked against the internal VersionMetadata LRU cache. If
+// it does not exist, it will be queried for via the snapshotDB and added to the
+// LRU cache. An error is returned if the VersionMetadata cannot be decoded or
+// queried for in the snapshotDB.
+func (ndb *nodeDB) GetVersionMetadata(version int64) (*VersionMetadata, error) {
+	key := metadataKeyFormat.Key(version)
+
+	v, ok := ndb.vmCache.Get(string(key))
+	if ok {
+		return v.(*VersionMetadata), nil
+	}
+
+	bz, err := ndb.snapshotDB.Get(key)
+	if err != nil {
+		return nil, err
+	}
+
+	// In cases where version metadata does not exist, say for a non-empty tree,
+	// we construct metadata on the fly. This is for backwards compatibility WRT
+	// to snapshotting.
+	if len(bz) == 0 {
+		return &VersionMetadata{
+			Version:  version,
+			Snapshot: ndb.opts.KeepEvery != 0 && version%ndb.opts.KeepEvery == 0,
+		}, nil
+	}
+
+	vm, err := unmarshalVersionMetadata(bz)
+	if err != nil {
+		return nil, err
+	}
+
+	_ = ndb.vmCache.Add(string(key), vm)
+	return vm, nil
+}
+
+// SetVersionMetadata saves a VersionMetadata in the snapshotDB and performs a
+// write-through to the VersionMetadata LRU cache. It returns an error if encoding
+// or saving to the snapshotDB fails.
+func (ndb *nodeDB) SetVersionMetadata(vm *VersionMetadata) error {
+	key := metadataKeyFormat.Key(vm.Version)
+
+	bz, err := vm.marshal()
+	if err != nil {
+		return err
+	}
+
+	if err := ndb.snapshotDB.Set(key, bz); err != nil {
+		return err
+	}
+
+	_ = ndb.vmCache.Add(string(key), vm)
+	return nil
+}
+
+// DeleteVersionMetadata removes a VersionMetadata object from the snapshotDB
+// and the LRU cache.
+func (ndb *nodeDB) DeleteVersionMetadata(version int64) error {
+	vm, err := ndb.GetVersionMetadata(version)
+	if err != nil {
+		return err
+	}
+
+	key := metadataKeyFormat.Key(vm.Version)
+
+	if err := ndb.snapshotDB.Delete(key); err != nil {
+		return err
+	}
+
+	ndb.vmCache.Remove(string(key))
+	return nil
 }
 
 func (ndb *nodeDB) isRecentVersion(version int64) bool {
@@ -201,9 +281,13 @@ func (ndb *nodeDB) HasSnapshot(hash []byte) (bool, error) {
 }
 
 // SaveTree takes a rootNode and version. Saves all nodes in tree using SaveBranch
-func (ndb *nodeDB) SaveTree(root *Node, version int64) []byte {
-	flushToDisk := ndb.isSnapshotVersion(version)
-	return ndb.SaveBranch(root, flushToDisk)
+func (ndb *nodeDB) SaveTree(root *Node, version int64) ([]byte, error) {
+	vm, err := ndb.GetVersionMetadata(version)
+	if err != nil {
+		return nil, err
+	}
+
+	return ndb.SaveBranch(root, vm.Snapshot), nil
 }
 
 // SaveBranch saves the given node and all of its descendants.
@@ -261,36 +345,49 @@ func (ndb *nodeDB) DeleteVersionFromRecent(version int64, checkLatestVersion boo
 
 func (ndb *nodeDB) deleteVersion(version int64, checkLatestVersion, memOnly bool) error {
 	if ndb.versionReaders[version] > 0 {
-		return errors.Errorf("unable to delete version %v, it has %v active readers",
-			version, ndb.versionReaders[version])
+		return errors.Errorf("unable to delete version %v, it has %v active readers", version, ndb.versionReaders[version])
 	}
 
-	err := ndb.deleteOrphans(version, memOnly)
+	vm, err := ndb.GetVersionMetadata(version)
 	if err != nil {
 		return err
 	}
-	ndb.deleteRoot(version, checkLatestVersion, memOnly)
+
+	if err := ndb.deleteOrphans(version, memOnly, vm.Snapshot); err != nil {
+		return err
+	}
+
+	ndb.deleteRoot(version, checkLatestVersion, memOnly, vm.Snapshot)
 	return nil
 }
 
 // Saves orphaned nodes to disk under a special prefix.
 // version: the new version being saved.
 // orphans: the orphan nodes created since version-1
-func (ndb *nodeDB) SaveOrphans(version int64, orphans map[string]int64) {
+func (ndb *nodeDB) SaveOrphans(version int64, orphans map[string]int64) error {
 	ndb.mtx.Lock()
 	defer ndb.mtx.Unlock()
 
 	toVersion := ndb.getPreviousVersion(version)
 
 	for hash, fromVersion := range orphans {
-		flushToDisk := false
+		var flushToDisk bool
+
+		vm, err := ndb.GetVersionMetadata(fromVersion)
+		if err != nil {
+			return err
+		}
+
 		if ndb.opts.KeepEvery != 0 {
 			// if snapshot version in between fromVersion and toVersion INCLUSIVE, then flush to disk.
-			flushToDisk = fromVersion/ndb.opts.KeepEvery != toVersion/ndb.opts.KeepEvery || ndb.isSnapshotVersion(fromVersion)
+			flushToDisk = fromVersion/ndb.opts.KeepEvery != toVersion/ndb.opts.KeepEvery || vm.Snapshot
 		}
+
 		debug("SAVEORPHAN %v-%v %X flushToDisk: %t\n", fromVersion, toVersion, hash, flushToDisk)
 		ndb.saveOrphan([]byte(hash), fromVersion, toVersion, flushToDisk)
 	}
+
+	return nil
 }
 
 // Saves a single orphan to recentDB. If flushToDisk, persist to disk as well.
@@ -314,17 +411,19 @@ func (ndb *nodeDB) saveOrphan(hash []byte, fromVersion, toVersion int64, flushTo
 
 // deleteOrphans deletes orphaned nodes from disk, and the associated orphan
 // entries.
-func (ndb *nodeDB) deleteOrphans(version int64, memOnly bool) error {
+func (ndb *nodeDB) deleteOrphans(version int64, memOnly, isSnapshot bool) error {
 	if ndb.opts.KeepRecent != 0 {
 		ndb.deleteOrphansMem(version)
 	}
-	if ndb.isSnapshotVersion(version) && !memOnly {
+
+	if isSnapshot && !memOnly {
 		predecessor := getPreviousVersionFromDB(version, ndb.snapshotDB)
 		traverseOrphansVersionFromDB(ndb.snapshotDB, version, func(key, hash []byte) {
 			ndb.snapshotBatch.Delete(key)
 			ndb.deleteOrphansHelper(ndb.snapshotDB, ndb.snapshotBatch, true, predecessor, key, hash)
 		})
 	}
+
 	return nil
 }
 
@@ -373,27 +472,33 @@ func (ndb *nodeDB) deleteOrphansHelper(db dbm.DB, batch dbm.Batch, flushToDisk b
 	}
 }
 
-func (ndb *nodeDB) PruneRecentVersions() (prunedVersions []int64, err error) {
+func (ndb *nodeDB) PruneRecentVersion() (int64, error) {
 	if ndb.opts.KeepRecent == 0 || ndb.latestVersion-ndb.opts.KeepRecent <= 0 {
-		return nil, nil
+		return 0, nil
 	}
+
 	ndb.mtx.Lock()
 	defer ndb.mtx.Unlock()
 
 	pruneVer := ndb.latestVersion - ndb.opts.KeepRecent
 	if ndb.versionReaders[pruneVer] > 0 {
-		return nil, nil
+		return 0, nil
 	}
-	err = ndb.deleteVersion(pruneVer, true, true)
+
+	if err := ndb.deleteVersion(pruneVer, true, true); err != nil {
+		return 0, err
+	}
+
+	vm, err := ndb.GetVersionMetadata(pruneVer)
 	if err != nil {
-		return nil, err
+		return 0, err
 	}
 
-	if ndb.isSnapshotVersion(pruneVer) {
-		return nil, nil
+	if vm.Snapshot {
+		return 0, nil
 	}
 
-	return append(prunedVersions, pruneVer), nil
+	return pruneVer, nil
 }
 
 func (ndb *nodeDB) nodeKey(hash []byte) []byte {
@@ -457,14 +562,15 @@ func getPreviousVersionFromDB(version int64, db dbm.DB) int64 {
 }
 
 // deleteRoot deletes the root entry from disk, but not the node it points to.
-func (ndb *nodeDB) deleteRoot(version int64, checkLatestVersion, memOnly bool) {
+func (ndb *nodeDB) deleteRoot(version int64, checkLatestVersion, memOnly, isSnapshot bool) {
 	if checkLatestVersion && version == ndb.getLatestVersion() {
 		panic("Tried to delete latest version")
 	}
 
 	key := ndb.rootKey(version)
 	ndb.recentBatch.Delete(key)
-	if ndb.isSnapshotVersion(version) && !memOnly {
+
+	if isSnapshot && !memOnly {
 		ndb.snapshotBatch.Delete(key)
 	}
 }
@@ -661,14 +767,25 @@ func (ndb *nodeDB) getRoots() (map[int64][]byte, error) {
 // loaded later.
 func (ndb *nodeDB) SaveRoot(root *Node, version int64) error {
 	if len(root.hash) == 0 {
-		panic("Hash should not be empty")
+		panic("SaveRoot: root hash should not be empty")
 	}
-	return ndb.saveRoot(root.hash, version, ndb.isSnapshotVersion(version))
+
+	vm, err := ndb.GetVersionMetadata(version)
+	if err != nil {
+		return err
+	}
+
+	return ndb.saveRoot(root.hash, version, vm.Snapshot)
 }
 
 // SaveEmptyRoot creates an entry on disk for an empty root.
 func (ndb *nodeDB) SaveEmptyRoot(version int64) error {
-	return ndb.saveRoot([]byte{}, version, ndb.isSnapshotVersion(version))
+	vm, err := ndb.GetVersionMetadata(version)
+	if err != nil {
+		return err
+	}
+
+	return ndb.saveRoot([]byte{}, version, vm.Snapshot)
 }
 
 func (ndb *nodeDB) saveRoot(hash []byte, version int64, flushToDisk bool) error {
@@ -838,6 +955,19 @@ func (ndb *nodeDB) traverseNodes(fn func(hash []byte, node *Node)) {
 	}
 }
 
+// restoreNodes restores nodes, which was orphaned, but after overwriting should not be orphans anymore
+func (ndb *nodeDB) restoreNodes(version int64) {
+	traverseOrphansVersionFromDB(ndb.recentDB, version, func(key, hash []byte) {
+		// Delete orphan key and reverse-lookup key.
+		ndb.recentBatch.Delete(key)
+	})
+
+	traverseOrphansVersionFromDB(ndb.snapshotDB, version, func(key, hash []byte) {
+		// Delete orphan key and reverse-lookup key.
+		ndb.snapshotBatch.Delete(key)
+	})
+}
+
 func (ndb *nodeDB) traverseNodesFromDB(db dbm.DB, fn func(hash []byte, node *Node)) {
 	nodes := []*Node{}
 
@@ -876,7 +1006,7 @@ func (ndb *nodeDB) String() string {
 	ndb.traverseNodes(func(hash []byte, node *Node) {
 		switch {
 		case len(hash) == 0:
-			str += fmt.Sprintf("<nil>\n")
+			str += "<nil>\n"
 		case node == nil:
 			str += fmt.Sprintf("%s%40x: <nil>\n", nodeKeyFormat.Prefix(), hash)
 		case node.value == nil && node.height > 0: